Setup Github Context
