+++ src/App.js
@@ -4,9 +4,11 @@ import About from './pages/About';
 import NotFound from './pages/NotFound';
 import Navbar from './components/layout/Navbar';
 import Footer from './components/layout/Footer';
+import { GithubProvider } from './context/github/GithubContext';
 
 const  App = () => {
   return (
+    <GithubProvider>
     <Router>
       <div className="flex flex-col justify-between h-screen">
         <Navbar />
@@ -23,6 +25,7 @@ const  App = () => {
         <Footer />
       </div>
     </Router>
+    </GithubProvider>
   );
 }
