+++ src/components/users/UserResults.jsx
@@ -1,27 +1,14 @@
-import { useEffect, useState } from "react";
+import { useEffect, useContext } from "react";
 import UserItem from './UserItem';
 import Spinner from "../layout/Spinner";
+import GithubContext from "../../context/github/GithubContext"
 
 const UserResults = () => {
-  const [users, setUsers] = useState([]);
-  const [loading, setLoading] = useState(true);
+  const {users, loading, fetchUsers } = useContext(GithubContext)
 
   useEffect(() => {
     fetchUsers()
   }, [])
-
-  const fetchUsers = async() => {
-    const response = await fetch(`${process.env.REACT_APP_GITHUB_URL}/users`, {
-      header: {
-        Authorization: `token ${process.env.REACT_APP_GITHUB_TOKEN}`
-      }
-    })
-
-    const data = await response.json()
-
-    setUsers(data);
-    setLoading(false);
-  }
 
   if(!loading) {
     return (
